+++ src/System.IO.FileSystem/tests/DirectoryInfo/GetSetAttributes.cs
@@ -6,75 +6,24 @@ using Xunit;
 
 namespace System.IO.Tests
 {
-    public class DirectoryInfo_GetSetAttributes : FileSystemTest
+    public class DirectoryInfo_GetSetAttributes : InfoGetSetAttributes<DirectoryInfo>
     {
-        #region Utilities
-
-        protected virtual FileAttributes Get(string path)
-        {
-            var info = new DirectoryInfo(path);
-            return info.Attributes;
-        }
-
-        protected virtual void Set(string path, FileAttributes attributes)
-        {
-            var info = new DirectoryInfo(path);
-            info.Attributes = attributes;
-        }
-
-        #endregion
-
-        [Fact]
-        public void NullParameters()
-        {
-            Assert.Throws<ArgumentNullException>(() => Get(null));
-            Assert.Throws<ArgumentNullException>(() => Set(null, FileAttributes.Normal));
-        }
-
-        [Fact]
-        public void InvalidParameters()
-        {
-            Assert.Throws<ArgumentException>(() => Get(string.Empty));
-            Assert.Throws<ArgumentException>(() => Set(string.Empty, FileAttributes.Normal));
-        }
-
-        // In NetFX we ignore "not found" errors, which leaves the attributes
-        // state as invalid (0xFFFFFFFF), which makes all flags true.
-
-        [Theory, MemberData(nameof(TrailingCharacters))]
-        public void GetAttributes_MissingFile(char trailingChar)
-        {
-            Assert.Equal((FileAttributes)(-1), Get(GetTestFilePath() + trailingChar));
-        }
-
-        [Theory, MemberData(nameof(TrailingCharacters))]
-        public void GetAttributes_MissingDirectory(char trailingChar)
-        {
-            Assert.Equal((FileAttributes)(-1), Get(Path.Combine(GetTestFilePath(), "file" + trailingChar)));
-        }
-
-        [Theory, MemberData(nameof(TrailingCharacters))]
-        public void SetAttributes_MissingFile(char trailingChar)
-        {
-            Assert.Throws<FileNotFoundException>(() => Set(GetTestFilePath() + trailingChar, FileAttributes.ReadOnly));
-        }
-
-        [Theory, MemberData(nameof(TrailingCharacters))]
-        public void SetAttributes_MissingDirectory(char trailingChar)
-        {
-            Assert.Throws<DirectoryNotFoundException>(() => Set(Path.Combine(GetTestFilePath(), "file" + trailingChar), FileAttributes.ReadOnly));
-        }
+        protected override FileAttributes Get(string path) => new DirectoryInfo(path).Attributes;
+        protected override void Set(string path, FileAttributes attributes) => new DirectoryInfo(path).Attributes = attributes;
+        protected override string CreateItem(string path = null) => Directory.CreateDirectory(path ?? GetTestFilePath()).FullName;
+        protected override DirectoryInfo CreateInfo(string path) => new DirectoryInfo(path);
+        protected override void DeleteItem(string path) => Directory.Delete(path);
+        protected override bool IsDirectory => true;
 
         [Theory]
         [InlineData(FileAttributes.ReadOnly)]
         [PlatformSpecific(TestPlatforms.AnyUnix)]  // Tests Unix file attributes
         public void UnixAttributeSetting(FileAttributes attr)
         {
-            var test = new DirectoryInfo(GetTestFilePath());
-            test.Create();
-            Set(test.FullName, attr);
-            Assert.Equal(attr | FileAttributes.Directory, Get(test.FullName));
-            Set(test.FullName, 0);
+            string path = CreateItem();
+            Set(path, attr);
+            Assert.Equal(attr | FileAttributes.Directory, Get(path));
+            Set(path, 0);
         }
 
         [Theory]
@@ -97,11 +46,10 @@ namespace System.IO.Tests
         [PlatformSpecific(TestPlatforms.Windows)]  // Tests Windows file attributes
         public void WindowsAttributeSetting(FileAttributes attr)
         {
-            var test = new DirectoryInfo(GetTestFilePath());
-            test.Create();
-            Set(test.FullName, attr);
-            Assert.Equal(attr | FileAttributes.Directory, Get(test.FullName));
-            Set(test.FullName, 0);
+            string path = CreateItem();
+            Set(path, attr);
+            Assert.Equal(attr | FileAttributes.Directory, Get(path));
+            Set(path, 0);
         }
 
         [Theory]
@@ -114,10 +62,9 @@ namespace System.IO.Tests
         [PlatformSpecific(TestPlatforms.AnyUnix)]  // Unix-invalid file attributes that don't get set
         public void UnixInvalidAttributes(FileAttributes attr)
         {
-            var path = GetTestFilePath();
-            File.Create(path).Dispose();
+            var path = CreateItem();
             Set(path, attr);
-            Assert.Equal(FileAttributes.Normal, Get(path));
+            Assert.Equal(FileAttributes.Directory, Get(path));
         }
 
         [Theory]
@@ -129,10 +76,9 @@ namespace System.IO.Tests
         [PlatformSpecific(TestPlatforms.Windows)]  // Windows-invalid file attributes that don't get set
         public void WindowsInvalidAttributes(FileAttributes attr)
         {
-            var path = GetTestFilePath();
-            File.Create(path).Dispose();
+            var path = CreateItem();
             Set(path, attr);
-            Assert.Equal(FileAttributes.Normal, Get(path));
+            Assert.Equal(FileAttributes.Directory, Get(path));
         }
 
         [Theory]
@@ -141,9 +87,8 @@ namespace System.IO.Tests
         [PlatformSpecific(TestPlatforms.AnyUnix)]  // Unix-invalid file attributes that throw
         public void UnixInvalidAttributes_ThrowArgumentException(FileAttributes attr)
         {
-            var test = new DirectoryInfo(GetTestFilePath());
-            test.Create();
-            Assert.Throws<ArgumentException>(() => Set(test.FullName, attr));
+            var path = CreateItem();
+            Assert.Throws<ArgumentException>(() => Set(path, attr));
         }
 
         [Theory]
@@ -153,43 +98,8 @@ namespace System.IO.Tests
         [PlatformSpecific(TestPlatforms.Windows)]  // Windows-invalid file attributes that throw
         public void WindowsInvalidAttributes_ThrowArgumentException(FileAttributes attr)
         {
-            var test = new DirectoryInfo(GetTestFilePath());
-            test.Create();
-            Assert.Throws<ArgumentException>(() => Set(test.FullName, attr));
-        }
-
-        [ConditionalFact(nameof(CanCreateSymbolicLinks))]
-        public void SymLinksAreReparsePoints()
-        {
-            var path = GetTestFilePath();
-            var linkPath = GetTestFilePath();
-
-            Directory.CreateDirectory(path);
-            Assert.True(MountHelper.CreateSymbolicLink(linkPath, path, isDirectory: true));
-
-            Assert.NotEqual(FileAttributes.ReparsePoint, FileAttributes.ReparsePoint & Get(path));
-            Assert.Equal(FileAttributes.ReparsePoint, FileAttributes.ReparsePoint & Get(linkPath));
-        }
-
-        [ConditionalFact(nameof(CanCreateSymbolicLinks))]
-        public void SymLinksReflectSymLinkAttributes()
-        {
-            var path = GetTestFilePath();
-            var linkPath = GetTestFilePath();
-
-            Directory.CreateDirectory(path);
-            Assert.True(MountHelper.CreateSymbolicLink(linkPath, path, isDirectory: true));
-
-            Set(path, FileAttributes.ReadOnly);
-            try
-            {
-                Assert.Equal(FileAttributes.ReadOnly, FileAttributes.ReadOnly & Get(path));
-                Assert.NotEqual(FileAttributes.ReadOnly, FileAttributes.ReadOnly & Get(linkPath));
-            }
-            finally
-            {
-                Set(path, Get(path) & ~FileAttributes.ReadOnly);
-            }
+            var path = CreateItem();
+            Assert.Throws<ArgumentException>(() => Set(path, attr));
         }
     }
 }
